fix(food): address Copilot review (PRD-124)
- Split hero-paths.ts into a browser-safe module (constants + URL +
  validators) and hero-paths.node.ts (node:path + env-reading abs-path
  helpers). HeroImageUploader's bundle no longer drags node:path or
  process.env into Vite.
- Import HERO_ALLOWED_MIME_TYPES from paths.ts in types.ts so the Zod
  schema module no longer reaches into the sharp/fs-bearing service.ts.
- Round up + clamp to 1 MB in the drop-zone size hint and the
  validation toast so sub-1MB caps don't render "0 MB".
